requirements.txt
langchain
langchain-groq
langchain-community
langchain-core
streamlit
python-dotenv